+++ CHANGELOG.md
@@ -104,4 +104,4 @@
 * Define `Persistable` instance for `Send (Text :&: Keyboard)`
 * Move modified versions of `Capacity` and `Silently` datatypes to `Message` module
 * Define `Persistable` instances for `Edit` and `Delete` messages
-*
+* Add `persist_` method in `Persistable` that not trying to decode JSON response