+++ CHANGELOG.md
@@ -106,3 +106,4 @@
 * Define `Persistable` instances for `Edit` and `Delete` messages
 * Add `persist_` method in `Persistable` that not trying to decode JSON response
 * Extract file ID from `File` datatypes to use these datatypes for sending as messages
+* Define `Persistable` instance for dropping `Notification`s on pressing inline buttons